se inician modulos de almacen

+++ database/migrations/2018_09_04_143834_create_permission_role_table.php
@@ -14,13 +14,10 @@ class CreatePermissionRoleTable extends Migration
     public function up()
     {
         Schema::create('permission_role', function (Blueprint $table) {
-            $table->unsignedSmallInteger('role_id');
+            $table->unsignedInteger('role_id');
             $table->string('permission_id', 40);
 
             $table->foreign('role_id')->references('id')->on('roles');
-            //$table->foreign('permission_id')->references('id')->on('permissions')->onUpdate('cascade');
-
-
         });
     }
